requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-appconf==1.0.3
django-ckeditor==5.9.0
django-countries==5.5
django-crispy-forms==1.8.1
django-filter==2.2.0
django-image-cropping==1.3.0
django-js-asset==1.2.2
easy-thumbnails==2.7
freeze==1.0.10
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.8
wrapt==1.11.2
